menu changes

+++ src/main/java/org/example/Test.java
@@ -4,7 +4,7 @@ import java.util.LinkedHashMap;
 
 public class Test {
 
-    public static String example(User user, LinkedHashMap<String, Float> returnFromBank) {
+    public static String example(User user, LinkedHashMap<String, Double> returnFromBank) {
         String result = "";
         for (String currency : user.getCurrency()) {
             result += "Курс в " + BankNames.valueOf(user.getBank()) + ": " + currency + "/UAH\n";
@@ -23,11 +23,11 @@ public class Test {
         User user = new User(1L);
         user.addCurrency("EUR");
 
-        LinkedHashMap<String, Float> returnFromBank = new LinkedHashMap<>();
-        returnFromBank.put("USD_buy", 24.5F);
-        returnFromBank.put("USD_sell", 25.5F);
-        returnFromBank.put("EUR_buy", 26.5F);
-        returnFromBank.put("EUR_sell", 27.5F);
+        LinkedHashMap<String, Double> returnFromBank = new LinkedHashMap<>();
+        returnFromBank.put("USD_buy", 24.5);
+        returnFromBank.put("USD_sell", 25.5);
+        returnFromBank.put("EUR_buy", 26.5);
+        returnFromBank.put("EUR_sell", 27.5);
 
         System.out.println(example(user,returnFromBank));
     }
